renamed script to period, changed minimize-function for significance criteria in order to allow for bounds, established assert_allclose to returned significances, removed Latex from plot
