+++ gammapy/time/tests/test_plot_periodogram.py
@@ -1,7 +1,7 @@
 import pytest
-from ..lomb_scargle import lomb_scargle
+from ..period import lomb_scargle
 from ..plot_periodogram import plot_periodogram
-from .test_lomb_scargle import simulate_test_data
+from .test_period import simulate_test_data
 
 
 @pytest.mark.parametrize('test_case', [
@@ -18,7 +18,7 @@ def test_lomb_scargle_plot(test_case):
         test_data['t'], test_data['y'], test_data['dy'], test_case['dt'],
         test_case['max_period'], test_case['criteria'], test_case['n_bootstraps'],
     )
-    plot_periodogram(
+    fig = plot_periodogram(
         test_data['t'], test_data['y'], test_data['dy'], result['pgrid'],
         result['psd'], result['swf'], result['period'],
         result['significance']
